Implement Un-eject
When ejecting a track, save a pointer to the ejected track.  If the user presses eject on an empty track in any deck or sampler, the saved track is loaded.

The track is loaded fresh, not as a "Deck clone" operation.

This enables a sort of "eject undo" as well as "track stashing".   It requires no changes to controller configs.

This required a refactor of the eject code to remove it from the EngineBuffer.  Tests still pass. Hand-testing works.